renamed common_VerifierAlg.ts to CommonVerifierAlg.ts and changed references in src/VerifierAlg
