+++ src/VerifierAlg/ES256VerifierAlg.ts
@@ -3,8 +3,7 @@ import { sha256, toEthereumAddress } from '../Digest'
 import type { VerificationMethod } from 'did-resolver'
 import { bytesToHex, EcdsaSignature } from '../util'
 import { verifyBlockchainAccountId } from '../blockchains'
-
-import * as common_VerifierAlg from './common_VerifierAlg'
+import * as CommonVerifierAlg from './CommonVerifierAlg'
 
 const secp256r1 = new EC('p256')
 
@@ -16,7 +15,7 @@ interface LegacyVerificationMethod extends VerificationMethod {
 
 export function verifyES256(data: string, signature: string, authenticators: VerificationMethod[]): VerificationMethod {
   const hash: Uint8Array = sha256(data)
-  const sigObj: EcdsaSignature = common_VerifierAlg.toSignatureObject(signature)
+  const sigObj: EcdsaSignature = CommonVerifierAlg.toSignatureObject(signature)
   const fullPublicKeys = authenticators.filter(({ ethereumAddress, blockchainAccountId }) => {
     return typeof ethereumAddress === 'undefined' && typeof blockchainAccountId === 'undefined'
   })
@@ -26,7 +25,7 @@ export function verifyES256(data: string, signature: string, authenticators: Ver
 
   let signer: VerificationMethod | undefined = fullPublicKeys.find((pk: VerificationMethod) => {
     try {
-      const pubBytes = common_VerifierAlg.extractPublicKeyBytes(pk)
+      const pubBytes = CommonVerifierAlg.extractPublicKeyBytes(pk)
       return secp256r1.keyFromPublic(pubBytes).verify(hash, <SignatureInput>sigObj)
     } catch (err) {
       return false
@@ -48,9 +47,9 @@ export function verifyRecoverableES256(
 ): VerificationMethod {
   let signatures: EcdsaSignature[]
   if (signature.length > 86) {
-    signatures = [common_VerifierAlg.toSignatureObject(signature, true)]
+    signatures = [CommonVerifierAlg.toSignatureObject(signature, true)]
   } else {
-    const so = common_VerifierAlg.toSignatureObject(signature, false)
+    const so = CommonVerifierAlg.toSignatureObject(signature, false)
     signatures = [
       { ...so, recoveryParam: 0 },
       { ...so, recoveryParam: 1 },
@@ -66,7 +65,7 @@ export function verifyRecoverableES256(
     const recoveredAddress: string = toEthereumAddress(recoveredPublicKeyHex)
 
     const signer: VerificationMethod | undefined = authenticators.find((pk: VerificationMethod) => {
-      const keyHex = bytesToHex(common_VerifierAlg.extractPublicKeyBytes(pk))
+      const keyHex = bytesToHex(CommonVerifierAlg.extractPublicKeyBytes(pk))
       return (
         keyHex === recoveredPublicKeyHex ||
         keyHex === recoveredCompressedPublicKeyHex ||
